new smg weapon
rapid fire of weak projectiles.
does 10% per projectile, but fires 10x faster than basic gun, so DPS is the same.

An attempt at a laser weapon became a machine gun

+++ BulletHell Dungeon Project/Assets/Scripts/NPC/HealthBar.cs
@@ -7,14 +7,14 @@ public class HealthBar : MonoBehaviour {
 
 	public Transform currentValTransform;
 
-	private int MaxHealth;
-	public int health;
+	private float MaxHealth;
+	public float health;
 
 	void Start(){
 		MaxHealth = health;
 	}
 
-	public void ChangeHealth(int val){
+	public void ChangeHealth(float val){
 		health = Mathf.Clamp (health + val, 0, MaxHealth);
 	}
